refactor: use double instead of float for money

+++ src/main/java/com/github/tareksaeed0/ExpirableProduct.java
@@ -5,7 +5,7 @@ import java.time.LocalDateTime;
 public class ExpirableProduct extends Product implements Expirable {
 	private LocalDateTime expirationTime;
 
-	public ExpirableProduct(String name, float price, int quantity,
+	public ExpirableProduct(String name, double price, int quantity,
 			LocalDateTime expirationTime) {
 		super(name, price, quantity);
